requirement.txt
streamlit==1.44.1
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
matplotlib=3.9.1
